- New api handles - Some code fixes - New README.md

+++ src/core/core.cpp
@@ -4,22 +4,25 @@
 #include <fstream>
 #include "core.hpp"
 
-bool KiViDbCore::Core::is_directory_exists(const char *path) {
-  struct stat info{};
-  if (stat(path, &info) != 0) {
-	return false;
-  } else if (info.st_mode & S_IFDIR) {
-	return true;
-  }
-  return false;
+// Check if the cluster exists
+bool KiViDbCore::Core::cluster_exists(const std::string &cluster_name) const {
+  // Check if the cluster exists
+  return std::filesystem::exists(db_folder_name + cluster_name);
 }
 
+// Check if the document exists
+bool KiViDbCore::Core::document_exists(const std::string &cluster_name, const std::string &document_name) const {
+  return std::filesystem::exists(db_folder_name + cluster_name + "/" + document_name);
+}
+
+// Create a new cluster
 void KiViDbCore::Core::create_cluster(const std::string &cluster_name) {
-  std::string cluster_path = db_folder_name + cluster_name;
-  if (!cluster_path.ends_with("/")) {
-	cluster_path += "/";
-  }
-  if (!is_directory_exists(cluster_path.c_str())) {
+  // Check if cluster already exists
+  if (!cluster_exists(cluster_name)) {
+	std::string cluster_path = db_folder_name + cluster_name;
+	if (!cluster_path.ends_with("/")) {
+	  cluster_path += "/";
+	}
 	std::filesystem::create_directory(cluster_path);
 	cluster_array.append(Cluster{cluster_name, cluster_path});
   }
@@ -27,9 +30,11 @@ void KiViDbCore::Core::create_cluster(const std::string &cluster_name) {
 
 // Get Cluster by name
 Cluster KiViDbCore::Core::get_cluster(const std::string &cluster_name) const {
-  for (int i = 0; i < cluster_array.size(); i++) {
-	if (cluster_array[i].name == cluster_name) {
-	  return cluster_array[i];
+  if (cluster_exists(cluster_name)) {
+	for (int i = 0; i < cluster_array.size(); i++) {
+	  if (cluster_array[i].name == cluster_name) {
+		return cluster_array[i];
+	  }
 	}
   }
   return Cluster{"", ""};
@@ -55,47 +60,80 @@ void KiViDbCore::Core::update_cluster_array() {
 	  std::string document_content;
 	  document_content_file >> document_content;
 	  // Create Document object
-	  _doc_vector[i++] = Document{document.path().filename(), document_content};
+	  _doc_vector[i++] = Document(document.path().filename(), document_content);
 	}
 	cluster_array.append(Cluster{cluster_name, cluster_path, _doc_vector});
   }
 }
-
+// Getting copy of cluster array
 ClusterArray KiViDbCore::Core::get_all_clusters() const {
   return cluster_array.copy();
 }
+
+// Delete a cluster by its name
 void KiViDbCore::Core::delete_cluster(const std::string &cluster_name) {
-  for (int i = 0; i < cluster_array.size(); i++) {
-	if (cluster_array[i].name == cluster_name) {
-	  std::filesystem::remove_all(cluster_array[i].path);
-	  cluster_array.remove(i);
-	  return;
+  if (cluster_exists(cluster_name)) {
+	for (int i = 0; i < cluster_array.size(); i++) {
+	  if (cluster_array[i].name == cluster_name) {
+		std::filesystem::remove_all(cluster_array[i].path);
+		cluster_array.remove(i);
+		return;
+	  }
 	}
   }
 }
+
+// Get the document by its name
 Document KiViDbCore::Core::get_document(const std::string &cluster_name, const std::string &document_name) const {
   std::string document_path = db_folder_name + cluster_name + "/" + document_name;
-  if (!is_directory_exists(document_path.c_str())) {
-	return Document{"", ""};
-  }
-  for (const auto &entry : std::filesystem::directory_iterator(db_folder_name + cluster_name + "/")) {
-	if (entry.path().filename() == document_name) {
-	  return Document{document_name, document_path};
+  if (document_exists(cluster_name, document_name)) {
+	for (const auto &entry : std::filesystem::directory_iterator(db_folder_name + cluster_name + "/")) {
+	  if (entry.path().filename() == document_name) {
+		// Read Document content
+		std::ifstream document_content_file(entry.path());
+		std::string document_content;
+		document_content_file >> document_content;
+		return Document{document_name, document_content};
+	  }
 	}
   }
   return Document{"", ""};
 }
-
+// Create a new document
 void KiViDbCore::Core::create_document(const std::string &cluster_name,
 									   const std::string &document_name,
 									   const std::string &document_content) {
   std::string document_path = db_folder_name + cluster_name + "/" + document_name;
-  if (!is_directory_exists(document_path.c_str())) {
-	std::filesystem::create_directory(document_path);
+  if (!document_exists(cluster_name, document_name)) {
+	// Create file
+	std::ofstream document_file(document_path);
+	document_file << document_content;
+	document_file.close();
+	// Adding document to document array for given cluster
+	this->get_cluster(cluster_name).add_document(Document{document_name, document_content});
   }
-  this->get_cluster(cluster_name).add_document(Document{document_name, document_content});
-
 }
-bool KiViDbCore::Core::cluster_exists(const std::string &cluster_name) const {
-  return is_directory_exists((db_folder_name + cluster_name).c_str());
+
+// Updating document content
+void KiViDbCore::Core::update_document(const std::string &cluster_name,
+									   const std::string &document_name,
+									   const std::string &document_content) {
+  if (document_exists(cluster_name, document_name)) {
+	std::string document_path = db_folder_name + cluster_name + "/" + document_name;
+	// Erase file content and write new one
+	std::ofstream file(document_path);
+	file << document_content;
+	file.close();            // Close the file
+	update_cluster_array(); // Update cluster array (document arrays too)
+  }
+}
+
+// Delete a document by its name
+void KiViDbCore::Core::delete_document(const std::string &cluster_name, const std::string &document_name) {
+  if (document_exists(cluster_name, document_name)) {
+	std::string document_path = db_folder_name + cluster_name + "/" + document_name;
+	std::filesystem::remove(document_path);
+	// Remove document from document array for given cluster
+	this->get_cluster(cluster_name).remove_document(document_name);
+  }
 }
